Rename FileType -> OutputType
