+++ Sources/swift-package-list/OutputType+ExpressibleByArgument.swift
@@ -1,5 +1,5 @@
 //
-//  FileType+ExpressibleByArgument.swift
+//  OutputType+ExpressibleByArgument.swift
 //  swift-package-list
 //
 //  Created by Felix Herrmann on 19.05.22.
@@ -8,4 +8,4 @@
 import ArgumentParser
 import SwiftPackageListCore
 
-extension FileType: ExpressibleByArgument { }
+extension OutputType: ExpressibleByArgument { }
